Init refactor
