+++ src/pages.bkp/development-info/index.js
@@ -1,0 +1,16 @@
+import React from 'react';
+import Layout from "../../components/layout"
+import { Container, Title, FlexList } from "../../shared/styled-shared"
+
+const DevelopmentInfo = () => {
+
+  return (
+    <Layout>
+      <Container>
+        <Title>Teste</Title>
+      </Container>
+    </Layout>
+  )
+}
+
+export default DevelopmentInfo;
