+++ src/pages.bkp/index.js
@@ -1,0 +1,6 @@
+import React from "react"
+import Home from "./home"
+
+const IndexPage = () => <Home />
+
+export default IndexPage
