--- src/pages/index.js
@@ -1,6 +1,0 @@
-import React from "react"
-import Home from "./home"
-
-const IndexPage = () => <Home />
-
-export default IndexPage
